Remove store.chain.transactionStatus
